refactor: 경매 등록 서비스 디렉토리 분리 (#CHZZ-35)
